Display details of problems in contests.
Fix some bugs.

+++ web/src/test/java/org/verwandlung/voj/web/mapper/ContestContestantMapperTest.java
@@ -114,7 +114,7 @@ public class ContestContestantMapperTest {
 
 		int score = firstContestant.getScore();
 		Assertions.assertEquals(200, score);
-		int time = firstContestant.getTime();
+		long time = firstContestant.getTime();
 		Assertions.assertEquals(60, time);
 	}
